Only add nonstandard port to Host header
Refs #20.

There are some servers (e.g. github pages with custom domain)
which will fail if you provide a port number with the Host
header. The RFC seems to claim that it's only needed if using
a nonstandard port -- so only add the port to the Host header
in that case.

Signed-off-by: Rick Elrod <rick@elrod.me>